+++ android/src/main/kotlin/com/jarvan/fluwx/handlers/FluwxShareHandler.kt
@@ -134,7 +134,7 @@ internal interface FluwxShareHandler : CoroutineScope {
                 sourceByteArray.isEmpty() -> {
                     WXImageObject()
                 }
-                sourceByteArray.size > 500 * 1024 -> {
+                else -> {
                     WXImageObject().apply {
                         if (supportFileProvider && targetHigherThanN) {
                             setImagePath(getFileContentUri(sourceByteArray.toCacheFile(context, sourceImage.suffix)))
@@ -146,9 +146,6 @@ internal interface FluwxShareHandler : CoroutineScope {
                             }
                         }
                     }
-                }
-                else -> {
-                    WXImageObject(sourceByteArray)
                 }
             }
             val msg = WXMediaMessage()
@@ -246,15 +243,18 @@ internal interface FluwxShareHandler : CoroutineScope {
             val sourceFile = WeChatFile.createWeChatFile(map, assetFileDescriptor)
 
             val sourceByteArray = sourceFile.readByteArray()
-//            if (supportFileProvider && targetHigherThanN) {
-//                wxFileObject.filePath = getFileContentUri(sourceByteArray.toCacheFile(context, sourceFile.suffix))
-//            } else {
-                if (ContextCompat.checkSelfPermission(context, Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED) {
-                    wxFileObject.filePath = sourceByteArray.toExternalCacheFile(context, sourceFile.suffix)?.absolutePath
+
+            wxFileObject.apply {
+                if (supportFileProvider && targetHigherThanN) {
+                    setFilePath(getFileContentUri(sourceByteArray.toCacheFile(context, sourceFile.suffix)))
                 } else {
-                    permissionHandler?.requestStoragePermission()
+                    if (ContextCompat.checkSelfPermission(context, Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED) {
+                        filePath = sourceByteArray.toExternalCacheFile(context, sourceFile.suffix)?.absolutePath
+                    } else {
+                        permissionHandler?.requestStoragePermission()
+                    }
                 }
-//            }
+            }
 
             msg.thumbData = readThumbnailByteArray(call)
             val req = SendMessageToWX.Req()
@@ -270,10 +270,15 @@ internal interface FluwxShareHandler : CoroutineScope {
 
     private suspend fun readThumbnailByteArray(call: MethodCall, length: Int = SHARE_IMAGE_THUMB_LENGTH): ByteArray? {
         val thumbnailMap: Map<String, Any>? = call.argument(keyThumbnail)
+        val compress:Boolean = call.argument("compressThumbnail")?:true
         return thumbnailMap?.run {
             val thumbnailImage = WeChatFile.createWeChatFile(thumbnailMap, assetFileDescriptor)
             val thumbnailImageIO = ImagesIOIml(thumbnailImage)
-            compressThumbnail(thumbnailImageIO, length)
+            if(compress){
+                compressThumbnail(thumbnailImageIO, length)
+            }else{
+                thumbnailImageIO.readByteArray()
+            }
         }
     }