+++ ios/Classes/FluwxPlugin.m
@@ -46,12 +46,16 @@ BOOL handleOpenURLByFluwx = YES;
         result(@([WXApi openWXApp]));
     } else if ([@"payWithFluwx" isEqualToString:call.method]) {
         [self handlePayment:call result:result];
+    } else if ([@"payWithHongKongWallet" isEqualToString:call.method]) {
+        [self handleHongKongWalletPayment:call result:result];
     } else if ([@"launchMiniProgram" isEqualToString:call.method]) {
         [self handleLaunchMiniProgram:call result:result];
     } else if ([@"subscribeMsg" isEqualToString:call.method]) {
         [self handleSubscribeWithCall:call result:result];
     } else if ([@"autoDeduct" isEqualToString:call.method]) {
         [self handleAutoDeductWithCall:call result:result];
+    }else if([@"authByPhoneLogin" isEqualToString:call.method]){
+        [_fluwxAuthHandler handleAuthByPhoneLogin:call result:result];
     } else if ([call.method hasPrefix:@"share"]) {
         [_fluwxShareHandler handleShare:call result:result];
     } else {
@@ -60,7 +64,9 @@ BOOL handleOpenURLByFluwx = YES;
 }
 
 - (void)registerApp:(FlutterMethodCall *)call result:(FlutterResult)result {
-    if (!call.arguments[@"iOS"]) {
+    NSNumber* doOnIOS =call.arguments[@"iOS"];
+
+    if (![doOnIOS boolValue]) {
         result(@NO);
         return;
     }
@@ -107,6 +113,19 @@ BOOL handleOpenURLByFluwx = YES;
                                 Sign:sign completion:^(BOOL done) {
                 result(@(done));
             }];
+}
+
+- (void)handleHongKongWalletPayment:(FlutterMethodCall *)call result:(FlutterResult)result {
+    NSString *partnerId = call.arguments[@"prepayId"];
+    
+    WXOpenBusinessWebViewReq *req = [[WXOpenBusinessWebViewReq alloc] init];
+    req.businessType = 1;
+    NSMutableDictionary *queryInfoDic = [NSMutableDictionary dictionary];
+    [queryInfoDic setObject:partnerId forKey:@"token"];
+    req.queryInfoDic = queryInfoDic;
+    [WXApi sendReq:req completion:^(BOOL done) {
+        result(@(done));
+    }];
 }
 
 - (void)handleLaunchMiniProgram:(FlutterMethodCall *)call result:(FlutterResult)result {
@@ -171,12 +190,9 @@ BOOL handleOpenURLByFluwx = YES;
     return [WXApi handleOpenURL:url delegate:[FluwxResponseHandler defaultManager]];
 }
 
-- (BOOL) application:(UIApplication *)application
-continueUserActivity:(NSUserActivity *)userActivity
-  restorationHandler:(void (^)(NSArray *))restorationHandler {
-    return [WXApi handleOpenUniversalLink:userActivity delegate:[FluwxResponseHandler defaultManager]];
+- (BOOL)application:(UIApplication *)application continueUserActivity:(NSUserActivity *)userActivity restorationHandler:(void (^)(NSArray * _Nonnull))restorationHandler{
+        return [WXApi handleOpenUniversalLink:userActivity delegate:[FluwxResponseHandler defaultManager]];
 }
-
 - (void)scene:(UIScene *)scene continueUserActivity:(NSUserActivity *)userActivity  API_AVAILABLE(ios(13.0)){
     [WXApi handleOpenUniversalLink:userActivity delegate:[FluwxResponseHandler defaultManager]];
 }